Major update - copy of work version
Trying to main an open and a work version of the code has proven to be
too difficult so from now on this version will be the same as what I use
at work. The only real difference this bring is that the entire
framework expects all of these CFCs to live in a subfolder called
"Components". So you either have to create one straight from the root of
where you intend to implement Selenium or make an Application-mapping to
a folder somewhere else.

There are too many fixes, optimizations and changes to mention so let's
just leave it at: stuff got better.