Syntax and style changes, new features and fixes
Biggest change is that I went through and updated the syntax and code
style to (hopefully) be more consistent across the board. Hints have
also been improved here and there.

Had to add or increase calls to sleep() in various places because the
speed was so fast that it sometimes wouldn't be able to target elements
before they were fully rendered. This is mostly an issue with pages that
use a lot of jQuery or CSS3 animation, which the wait-method can't
identify and wait for to finish.

Element.cfc now has an getElement() method which allow you to search for
elements nested inside of that element. In addition I added made changes
to click() that should hopefully better catch unclickable actions and
give you better feedback.

textEquals() in ElementLocator.cfc has been changed a bit to try and
compensate for leading and trailing whitespace. Since this method relies
on XPath and I am not super experienced using that, I can't guarantee
that it works perfectly well.