Locator re-write
Too many changes to mention almost. The primary change is that Selenium's By-class - which is used to fetch elements - used to be abstracted away inside of fetchHTMLElements() in Browser.cfc. Now however I changed it so that it is its own separate object, called Locator.cfc.

This was required, and just overall smarter, as it allows us to use locators for things such as Selenium's ExpectedConditions-class which works on locators, as well as elements.

You can create Locator.cfc instances yourself OR you can use createLocator() in Browser.cfc to spawn it instead, which is the intended design.

On top of that numerous fixes and other changes have been made. Another important one is that getElements() from Browser.cfc has been removed and getElement() now has a "multiple"-argument for fetching multiple elements, which makes it similar in use to the ElementLocator.cfc methods.

getElement() now accepts a locator as argument while the ElementLocator.cfc-methods remain largely unchanged, except for the fact that you can pass an argument called "getLocator" to return the Locator.cfc-instance that is used by the method.