Fixes
Added code to clean tabs, carriage returns, new lines and extra spaces to getTextContent() since I am guessing most people don't want that.

Updated a few hints and added code to hopefully catch issues with select- and deselectByIndex() when the select-tag is empty (having no options or 1 empty option).

Fixed a path replacement issue for the absolute path to the webdriver binaries in WebDriverManager to make it Linux compatible