Optimization and cleaning
By the looks of it there's a lot going on here. For the most part it's optimization. Been trying to clean house and remove a lot of jank and cruft. In particular I removed a lot of private setters and getters as they were ultimately redundant.
Most of the private variables are set on init and they are checked via the cfargument anyway. Only a few private setters remain in place that provide extra checks - primarily for Java-objects. Private getters are similarly useless and don't provide anything outside of an extra function call. Having removed these and replaced them with direct references to the variables has overall increased the performance (nothing dramatic, but noticable in the long run).

As far as actual changes and new features go:

- Finally removed the hardcoded reference to the application-scope in WebDriverManager, making the pathToWebDriverBIN-argument required.

- The SelectElement no longer requires a reference to Browser.cfc and the Java WebElement. It now just requires the wrapping Element.cfc it belongs to. It utilizes the public methods to get the WebElement and Browser from its enclosing Element instead.

- Element.cfc has two new methods, both designed to scroll the element into view. One is the official/proper way to do it and the other is an alternative.

- Added the ability via an argument in Element.click() to not implicitly convert the text-argument's array indices to strings. Also added a catch and new exception message for when this argument is false and one of the array indices is not a string.

- Made it optional via an argument in Element.getTextContent() to have it remove extra tabs, spaces and newlines.

- Implemented a custom method in Browser.cfc for checking if an element is in a state that it can/should be fetched. This method is more forgiving than Selenium's isDisplayed(). Selenium's criteria is considered "strict" and which method is used can be controlled with a new method called useStrictVisibilityCheck(). The strict method is enabled by default.