Direct use of internal private vars
Not much going on really. Mostly small fixes and optimizations.

Biggest change is the addition of console and performance logs for the webdriver. This currently only works for Chrome, and was implemented as a way to primarily get network traces (what files were hit) which can be done using performance logs.

It was to sidestep having to implement a third party proxy in order to analyze network traffic. Might be useful for other people as well.