Many changes, Lucee-only
This project keeps growing and as a result I have found it increasingly hard to make it work for us internally while also make it usable for the public.

So as of this release this project is just here for those interested in seeing how using Selenium with Coldfusion can be done.

The documentation is out of date, internally we've moved to using Lucee instead of Adobe's version and no one has (as far as I can tell) ever downloaded and used this project anyway, so all in all I am not too troubled about breaking it for anyone else.