getTextContent, and a bug
Biggest change is really getTextContent() in Element.cfc that has been a pain and now should be much better, using either textContent or innerText to leverage the browser's ability to remove unwanted characters.

Finally got around to fixing the pathToWebDriverBIN-argument in WebDriverManager.cfc which would grab stuff from the Application-scope if not passed.