Object factory, removed event manager, fixes
The biggest chance is the addition of SeleniumObjectFactory. I had trouble getting the Selenium jars loaded with the latest version of Lucee, despite my attempts at converting them to valid OSGi bundles, and constantly had conflicts or were unable to find certain classes.

Rather than add even more arguments and add even if-statements to account for other methods of loading the jars I decided to abstract it all via a factory-class. It supports multiple ways of creating the Selenium objects, such as via the createObject bundle-name argument which accepts a folder where the jar files live. I kept the option for the Javaloader as well.

Aside from that I deleted the EventManager as I wasn't happy with the way it worked and couldn't really see a proper use for it going forward. It was an experiment anyway.

SelectElement still had a few places that referred to Components. as the hardcoded path for instantiating components, so those have been removed. Aside from that it's mostly small fixes, syntax changes, spelling corrections etc.