New stuff! Well, a different version of doing something existing really.
I was never that happy with Browser.waitUntil() and I wanted to not only do it differently, but also try making my own version as I've had some weird issues with Selenium's wait mechanisms.

So taking inspiration from FlaUI's Retry-class I've made a dedicated Wait-class. It's been tested but not extensively, and currently everything except for waitUntilElementLocatedByIsPresent is returning booleans instead of the actual element.

I haven't decided how to deal with this, or if I even want the wait methods to return the element you are waiting on. The reason for this is primarily that I use the Java-classes for speed and you'd get the raw Java WebElements back instead of an Element-instance.